+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2026-03-24 - Improvement: Remove the 25 images limit for login background images, resolves #815
 * 2026-03-18 - Bugfix: Add an option to harden non-editable block regions by introducing a custom block manager, resolves #472
 * 2026-03-18 - Feature: Smart menus: Add dynamic course setting to filter starred courses, resolves #446
 * 2026-03-16 - Improvement: Add placeholders to static menu items' title and URLs, resolves #423